requirements.txt
python-telegram-bot>=21.6
python-dotenv>=1.0.0
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
